Fixing tests 1-5
for test 1, in item.dart, I changed the substring from (0,2) to (0,1) so the abbreviation was 1 letter, not two.

In test 2, I changed in to_do_items.dart the return completed. It was not returning when complete because the color being checked was not the same color as the text.

For test 3, in to_do_items.dart, I changed in Widget build item.name to item.abbrev() so that the text would print the abbreviated name, not the full name.

For step 4, I fixed in to_do_dialog.dart an issue in inconsistent naming. I changed "OkButton" to "OKButton" so the names were consistent.

For step 5, I changed in handlenewitem from             Item item = const Item(name: "itemText");
 to  Item item =  Item(name: itemText); because it needs to refer to the actual ItemText. Also, in to_do_dialog.dart, I changed the arangment of CancelButton and OKButton because they were in each others command.